fatorVencimento: cast int (Carbon 3 float + PHP 8.4 deprecated)

+++ src/Util.php
@@ -528,7 +528,9 @@ final class Util
     public static function fatorVencimento($date, $format = 'Y-m-d')
     {
         $date = ($date instanceof Carbon) ? $date : Carbon::createFromFormat($format, $date)->setTime(0, 0, 0);
-        $fator = (new Carbon('1997-10-07'))->diffInDays($date);
+        // (int): Carbon 3 devolve float em diffInDays e o PHP 8.4 avisa
+        // deprecated na conversão implícita (fork Prodo).
+        $fator = (int) (new Carbon('1997-10-07'))->diffInDays($date);
         $limit = $fator % 9000;
         if ($limit >= 1000) {
             return $limit;
